services/wallet_core/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
redis==5.0.7
celery==5.4.0
django-celery-results==2.5.1
python-dotenv==1.0.1
requests==2.31.0
gunicorn>=21.2
redis>=5.0.0